+++ src/components/DebtSummaryPage/GoalPic.jsx
@@ -1,0 +1,12 @@
+function GoalPic() {
+  return (
+         <div className="bg-white h-full w-7/15 min-w-[250px] rounded-4xl shadow-[5px_5px_5px_1px_rgba(0,_0,_0,_0.1)]">
+            <img
+              src="./osakaCastle.jpg"
+              alt="Osaka Castle"
+              className="h-full w-full object-cover rounded-4xl"
+            />
+          </div>
+  )
+}
+export default GoalPic
